handled redirection when the  user cancel the journey in CUIRA
